+++ ports/kiit-codes-ts/src/formats/index.ts
@@ -1,5 +1,4 @@
-export * from "./catalog.js";
 export * from "./error-item.js";
 export * from "./problem.js";
 export * from "./code-detail.js";
-export * from "./codes-to-problem.js";
+export * from "./problem-converter.js";